# packages/api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: Another Society API
  version: 0.3.1
  description: >
    Foundation API contract for the Another Society v0.3 architecture. The server is authoritative for identity,
    ownership, land, inventory, economy, canonical layouts, stays, and gifts.


    Ordinary Visit is not globally live presence; choosing Stay commits public foreign presence. WORLD soft currency is
    server-issued/earned and not directly funded by real-money purchases in the initial business model.


    v0.3.1 is additive only (new optional response fields, three read endpoints). Every mutation that can move
    ownership/money/presence takes Idempotency-Key; a replay returns the original body with header `Idempotent-Replayed:
    true`. Stable error codes: unauthorized, forbidden, not_found, validation_error, conflict, revision_conflict,
    idempotency_key_required, idempotency_mismatch, plot_unavailable, reservation_invalid, already_has_home,
    insufficient_funds, stay_not_allowed, stay_capacity_full, not_staying, gift_not_allowed, rate_limited,
    account_inactive, internal_error.
servers:
- url: https://api.example.com/v1
  description: Production placeholder; replace during deployment.
...
paths:
...
  /cities/{cityId}/chunks:
    get:
      tags:
      - World
      operationId: getCityChunks
      parameters:
      - $ref: "#/components/parameters/CityId"
      - in: query
        name: minX
        required: true
        schema:
          type: integer
      - in: query
        name: minY
        required: true
        schema:
          type: integer
      - in: query
        name: maxX
        required: true
        schema:
          type: integer
      - in: query
        name: maxY
        required: true
        schema:
          type: integer
      responses:
        "200":
          description: Dynamic state for requested chunk window only.
          content:
            application/json:
              schema:
                type: object
                required:
                - chunks
                properties:
                  chunks:
                    type: array
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/ChunkSummary"
        "422":
          $ref: "#/components/responses/ValidationError"
  /onboarding/plots:
    get:
      tags:
      - World
      operationId: getClaimablePlots
      parameters:
      - in: query
        name: city_id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: All currently active claimable plots in the onboarding city scope.
          content:
            application/json:
              schema:
                type: object
                required:
                - plots
                properties:
                  plots:
                    type: array
                    items:
                      $ref: "#/components/schemas/PlotSummary"
  /plots/{plotId}/reserve:
    post:
      tags:
      - World
      operationId: reservePlot
      parameters:
      - $ref: "#/components/parameters/PlotId"
      - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "201":
          description: Short-lived reservation created.
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/MutationResponse"
                - type: object
                  properties:
                    reservation_id:
                      type: string
                      format: uuid
                    expires_at:
                      type: string
                      format: date-time
        "409":
          $ref: "#/components/responses/Conflict"
  /plots/{plotId}/claim:
    post:
      tags:
      - World
      operationId: claimPlot
      parameters:
      - $ref: "#/components/parameters/PlotId"
      - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - reservation_id
              - structure_asset_id
              properties:
                reservation_id:
                  type: string
                  format: uuid
                structure_asset_id:
                  type: string
                structure_x_u:
                  type: integer
                  default: 0
                structure_y_u:
                  type: integer
                  default: 0
                structure_rot_q:
                  type: integer
                  minimum: 0
                  maximum: 3
      responses:
        "201":
          description: Plot claimed and starter property created atomically.
          content:
            application/json:
              schema:
                allOf:
                - $ref: "#/components/schemas/MutationResponse"
                - type: object
                  properties:
                    property:
                      $ref: "#/components/schemas/Property"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /cities/{cityId}/roads:
    get:
      tags:
      - World
      operationId: getCityRoads
      description: Active public road graph and civic anchors for client-side routing (A*). ETag is keyed to the city
        activation revision; send If-None-Match to get 304.
      parameters:
      - $ref: "#/components/parameters/CityId"
      - in: header
        name: If-None-Match
        schema:
          type: string
      responses:
        "200":
          description: Active road graph.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoadGraph"
        "304":
          description: Unchanged since the given ETag.
        "404":
          $ref: "#/components/responses/NotFound"
  /wander:
    get:
      tags:
      - Social
      operationId: wander
      description: A few publicly enterable homes, nearest-first with jitter. Not a ranking; never includes the caller,
        blocked residents, or closed homes.
      parameters:
      - in: query
        name: city_id
        required: true
        schema: &a2
          type: string
          format: uuid
      - in: query
        name: x_u
        schema: &a1
          type: integer
      - in: query
        name: y_u
        schema: *a1
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 10
          default: 5
      responses:
        "200":
          description: Wander candidates.
          content:
            application/json:
              schema:
                type: object
                required:
                - properties
                properties:
                  properties:
                    type: array
                    items:
                      allOf:
                      - $ref: "#/components/schemas/PropertyPreview"
                      - type: object
                        properties:
                          center_x_u: *a1
                          center_y_u: *a1
                          frontage_node_id:
                            type: string
                            format: uuid
        "422":
          $ref: "#/components/responses/ValidationError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
      description: Scoped to (user, operation). Same key + same request replays the stored response; same key + different
        request is 409 idempotency_mismatch.
    CityId:
      in: path
      name: cityId
      required: true
      schema:
        type: string
        format: uuid
    PlotId:
      in: path
      name: plotId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found or intentionally hidden.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: State conflict, duplicate/race, stale revision, or idempotency mismatch.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ValidationError:
      description: Request failed validation/business rules.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ChunkSummary:
      type: object
      required:
      - id
      - x
      - y
      - revision
      - plots
      - properties
      properties:
        id:
          type: string
          format: uuid
        x:
          type: integer
        y:
          type: integer
        revision:
          type: integer
          format: int64
        plots:
          type: array
          items:
            $ref: "#/components/schemas/PlotSummary"
        properties:
          type: array
          items:
            $ref: "#/components/schemas/PropertyPreview"
    PlotSummary:
      type: object
      required:
      - id
      - status
      - center_x_u
      - center_y_u
      - terrain_type
      - zone_type
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - latent
          - vacant
          - reserved
          - occupied
          - retired
        center_x_u:
          type: integer
        center_y_u:
          type: integer
        terrain_type:
          type: string
        zone_type:
          type: string
        scenic_tags:
          type: array
          items:
            type: string
        build_bounds:
          type: object
          additionalProperties: true
        chunk_x: *a1
        chunk_y: *a1
        labels:
          type: array
          description: Quiet map facts, never a recommendation score.
          items:
            type: string
            enum:
            - closer_to_town
            - near_park
            - near_neighbors
            - quieter_edge
            - forest_side
        reserved_by_me:
          type: boolean
        frontage_node_id:
          type: string
          format: uuid
          description: Public road node this plot fronts; route here for travel animation.
...
    Property:
      allOf:
      - $ref: "#/components/schemas/PropertyPreview"
      - type: object
        required:
        - owner_resident_id
        - allow_stays
        - stay_capacity
        - revision
        properties:
          owner_resident_id:
            type: string
            format: uuid
          allow_stays:
            type: boolean
          stay_capacity:
            type: integer
          is_primary_home:
            type: boolean
          revision:
            type: integer
            format: int64
          property_type:
            type: string
            enum:
            - home
            - studio
            - shop
            - other
          owner:
            type: object
            properties:
              id: *a2
              display_name:
                type: string
          space_ids:
            type: array
            items: *a2
          viewer:
            type: object
            properties:
              is_owner:
                type: boolean
              can_enter:
                type: boolean
              can_stay:
                type: boolean
          owner_away:
            type: boolean
            description: "Owner-only: true while the owner has an active foreign Stay."
...
    MutationResponse:
      type: object
      required:
      - request_id
      properties:
        request_id:
          type: string
        revision:
          type: integer
          format: int64
        resource_id:
          type: string
          format: uuid
...
security:
- bearerAuth: []